requirements.txt
Django==1.7.1
DjangoSEO==1.0
Fabric==1.10.1
Pillow==2.6.2
PyMySQL==0.6.3
South==1.0.1
Whoosh==2.5.7
django-admin-bootstrapped==2.3.0
django-debug-toolbar==1.2.2
django-djconfig==0.3.2
django-extensions==1.4.9
django-haystack==2.3.1
django-infinite-scroll-pagination==0.1.3
django-model-utils==2.2
django-robots==1.0
django-spirit==0.2.1
django-suit==0.2.12
django-taggit==0.13.0
django-wysiwyg==0.7.0
ecdsa==0.11
flake8==2.4.0
mccabe==0.3
mistune==0.3.1
mysql-connector-python==2.0.2
paramiko==1.15.2
pep8==1.5.7
pycrypto==2.6.1
pyflakes==0.8.1
pytz==2015.2
six==1.9.0
sqlparse==0.1.14
